rag/requirements.txt
# RAG Implementation Requirements with LangChain and Gemini Flash
langchain>=0.1.0
langchain-google-genai>=0.0.6
langchain-community>=0.0.10
langchain-chroma>=0.1.2
fastapi>=0.104.1
uvicorn>=0.24.0
chromadb>=0.4.18
google-generativeai>=0.3.2
python-dotenv>=1.0.0
pydantic>=2.5.0
python-multipart>=0.0.6
aiofiles>=23.2.1
httpx>=0.25.2
rich>=13.7.0
pytest>=7.4.3
pytest-asyncio>=0.21.1
numpy
pandas
